structuring fix
