+++ controllers/contacts/getAll.js
@@ -1,0 +1,12 @@
+const {listContacts} = require('../../service/index')
+
+const controlGetAll = async (req, res, next) => {
+  try {
+    const result = await listContacts()
+    res.json(result)
+  } catch (error) {
+    next(error)
+  }
+}
+
+module.exports = controlGetAll
